% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildsite.R
\name{buildsite}
\alias{buildsite}
\title{build and publish a website}
\usage{
buildsite(message = "update")
}
\arguments{
\item{message}{character string containing the message of the git commit.}
}
\description{
It uses \code{\link[blogdown]{build_site}} to build the site. It uses
\code{\link[mcdev]{publishdir}} to retrieve the directory of the website. And
it uses git commands to (i) move to the website directory, (ii) stage all new
changes, (iii) commit these change (together with the input argument) and (iv)
push upstream (which is publishing online).
}
\author{
Marc Choisy
}
