% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download}
\alias{download}
\title{Download a package file from an URL}
\usage{
download(url, package, file)
}
\arguments{
\item{url}{character string containing the URL to download from.}

\item{package}{character string containing the name of the package into which
to put the downloaded file.}

\item{file}{character string containing the name of the downloaded file
should be given.}
}
\description{
In case a raw data file is too big to be contained in a package, a solution
consists in downloading that file after package installation when the file is
needed for the first time. This is what the download function allows to do.
It will downlaod the file from the provided URL and put it in the extdata
directory of the provided package hierarchy, naming it using the name
provided as an argument.
}
\author{
Marc Choisy
}
