% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_site.R
\name{upload_site}
\alias{upload_site}
\title{Upload a website}
\usage{
upload_site(message = "update")
}
\arguments{
\item{message}{character string containing the message of the git commit.}
}
\description{
It uses \code{\link[mcdev]{publishdir}} to retrieve the directory of the
website, then, from the website directory, it uses git commands to (i) stage
all new changes, (ii) commit these change (together with the input argument
message), and (iii) push upstream (which is uploading).
}
\details{
The combination of `pkgdown::build_site` then `mcdev::upload_site` is
equivalent to `mcdev::buildsite` but we recommand to use the former option
instead.
}
\author{
Marc Choisy
}
